+++ packages/node/.resources.mjs
@@ -1,22 +1,7 @@
 export default [
   {
     path: '/models/',
-    source: '../../bundle/models/small',
-    mime: 'application/octet-steam'
-  },
-  {
-    path: '/models/',
-    source: '../../bundle/models/medium',
-    mime: 'application/octet-steam'
-  },
-  {
-    path: '/models/',
-    source: '../../bundle/models/large',
-    mime: 'application/octet-steam'
-  },
-  {
-    path: '/models/',
-    source: '../../bundle/models/modnet*',
+    source: '../../bundle/models/*',
     mime: 'application/octet-steam'
   }
 ];