+++ packages/web-data/.resources.mjs
@@ -6,22 +6,7 @@ export default [
   },
   {
     path: '/models/',
-    source: '../../bundle/models/small',
-    mime: 'application/octet-steam'
-  },
-  {
-    path: '/models/',
-    source: '../../bundle/models/medium',
-    mime: 'application/octet-steam'
-  },
-  // {
-  //   path: '/models/',
-  //   source: '../../bundle/models/large',
-  //   mime: 'application/octet-steam'
-  // },
-  {
-    path: '/models/',
-    source: '../../bundle/models/modnet*',
+    source: '../../bundle/models/*',
     mime: 'application/octet-steam'
   }
 ];