+++ packages/web/src/schema.ts
@@ -41,7 +41,18 @@ const ConfigSchema = z
       .describe('Progress callback.')
       .optional(),
     model: z
-      .enum(['small', 'medium', 'modnet', 'modnet_fp16'])
+      .preprocess((val) => {
+        switch (val) {
+          case 'large':
+            return 'isnet';
+          case 'small':
+            return 'isnet_quint8';
+          case 'medium':
+            return 'isnet_fp16';
+          default:
+            return val;
+        }
+      }, z.enum(['isnet', 'isnet_fp16', 'isnet_quint8', 'modnet', 'modnet_fp16' /*, 'modnet_quint8'*/]))
       .default('medium'),
     output: z
       .object({
@@ -58,25 +69,45 @@ const ConfigSchema = z
       })
       .default({})
   })
-  .default({});
+  .default({})
+  .transform((config) => {
+    if (config.debug) console.log('Config:', config);
+    if (config.debug && !config.progress) {
+      config.progress =
+        config.progress ??
+        ((key, current, total) => {
+          console.debug(`Downloading ${key}: ${current} of ${total}`);
+        });
+
+      if (!crossOriginIsolated) {
+        if (config.debug)
+          console.debug(
+            'Cross-Origin-Isolated is not enabled. Performance will be degraded. Please see  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer.'
+          );
+      }
+    }
+    if (config.device == 'gpu') {
+      switch (config.model) {
+        case 'isnet':
+          break;
+        case 'isnet_fp16':
+        case 'isnet_quint8':
+          if (config.debug) console.debug('Switching to f32 model for GPU');
+          config.model = 'isnet';
+          break;
+        case 'modnet':
+          break;
+        case 'modnet_fp16':
+          if (config.debug) console.debug('Switching to f32 model for GPU');
+          config.model = 'modnet';
+          break;
+      }
+    }
+    return config;
+  });
 
 type Config = z.infer<typeof ConfigSchema>;
 
 function validateConfig(configuration?: Config): Config {
-  const config = ConfigSchema.parse(configuration ?? {});
-  if (config.debug) console.log('Config:', config);
-  if (config.debug && !config.progress) {
-    config.progress =
-      config.progress ??
-      ((key, current, total) => {
-        console.debug(`Downloading ${key}: ${current} of ${total}`);
-      });
-
-    if (!crossOriginIsolated) {
-      console.debug(
-        'Cross-Origin-Isolated is not enabled. Performance will be degraded. Please see  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer.'
-      );
-    }
-  }
-  return config;
+  return ConfigSchema.parse(configuration ?? {});
 }