+++ packages/node/CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4.5]
+## [Unreleased]
 
 ### Added
 
-- Added ThirdPartyLicenses.json Added
+- Added Modnet models Added
+
+- Added isnet model for webgpu Added
+
+## [1.4.5]
+
+Added ThirdPartyLicenses.json Added
 
 ## [1.4.0]